rearrange binding registry building tests for TestRunnerManager
